+++ tasks/optimize.js
@@ -59,11 +59,27 @@ const wOpzer = async function (ctx, iEnv, configPath) {
     )
   }
 
+  // yarn 安装检查
+  if (config.yarn) {
+    const yarnVersion = await extOs.getYarnVersion()
+    if (yarnVersion) {
+      log(
+        'msg',
+        'info',
+        `${LANG.OPTIMIZE.YARN_VERSION}: ${chalk.green(yarnVersion)}`
+      )
+    } else {
+      throw new Error(
+        `${LANG.OPTIMIZE.INSTALL_YARN}: ${chalk.yellow('npm i yarn -g')}`
+      )
+    }
+  }
+
   const opzer = await seed.optimize({
     config,
     iEnv,
     ctx,
-    root: path.dirname(configPath),
+    root: path.dirname(configPath)
   })
 
   // handle exists check
@@ -91,11 +107,11 @@ const wOpzer = async function (ctx, iEnv, configPath) {
     const wServer = require('./server.js')
     const op = {
       livereload: opzer.ignoreLiveReload && !iEnv.livereload ? false : true,
-      ignoreServer: opzer.ignoreServer,
+      ignoreServer: opzer.ignoreServer
     }
 
     let afterConfig = await wServer.start(config, iEnv, op, {
-      appWillMount: opzer.appWillMount,
+      appWillMount: opzer.appWillMount
     })
     if (afterConfig) {
       config = afterConfig
@@ -141,10 +157,10 @@ const wOpzer = async function (ctx, iEnv, configPath) {
               r.push(item)
             })
             return r
-          })(),
+          })()
         })
         log('msg', 'success', [
-          `${LANG.OPTIMIZE.PRINT_HOME_PAGE}: ${chalk.yellow.bold(homePage)}`,
+          `${LANG.OPTIMIZE.PRINT_HOME_PAGE}: ${chalk.yellow.bold(homePage)}`
         ])
         // 第一次构建 打开 对应页面
         if (IS_WATCH && !isUpdate && !iEnv.silent && iEnv.proxy) {