+++ history.md
@@ -1,4 +1,20 @@
 # 版本信息
+## 3.11.1 (2020-04-21)
+* feat: 升级 `yyl-seed-webpack@2.5.2`
+* fix: 修复 默认端口 `5000` 不能用问题
+* fix: 修复配置 `--port 4000` 时 hmr 依然指向 5000 的问题
+
+## 3.11.0 (2020-04-21)
+* feat: 升级 `yyl-os@0.11.0`
+* feat: 升级 `yyl-hander@0.10.0`
+* feat: 升级 `yyl-seed-webpack@2.5.0`
+* feat: 升级 `yyl-server@0.3.8`
+* feat: yyl 支持 `yarn` 配置项
+* feat: 新增 `config.babelLoaderIncludes` 配置项
+* fix: 允许通过 `--port` 自定义 本地server 端口的功能
+* fix: 允许通过 `--proxy false` 自定义 是否打开 反向代理功能
+* fix: 修复快捷命令 `yyl r`, `yyl o`, `yyl d` 里面的参数 会覆盖 后面输入的 命令问题
+
 ## 3.10.2 (2020-04-01)
 * feat: 升级 `yyl-seed-webpack@2.3.1`
 * fix: 修复 `webpack` 项目启动 webpack-dev-server 时会出现接口占用误判断问题